PXF Docker environment for tests running
The goal of this patch is to begin automation of PXF testing. It contains
scripts and images for Docker environment, like it's done for other ADB
components and extensions. Existing upstream concourse scripts were reused where
it was possible. For now, we don't focus on hadoop integration testing, but do
unit testing, fdw and external-table regression testing.