Close connection to PXF when query to external table is canceled
The C-part of the PXF releases the context (cleanup_context) only on
the last call in the pxfprotocol_export and pxfprotocol_import functions.
That is, on errors, the context is not released, including
curl-connections to the Java-part are not closed.
Therefore, I added a callback to release resources on errors, which
releases the context, including closing curl-connections.