Refactor closing connection to pxf external-table and fdw
The C-part of the PXF external table releases the context
(cleanup_context) only on the last call in the pxfprotocol_export
and pxfprotocol_import functions.
The C-part of the PXF fdw releases the context (PxfBridgeCleanup)
only in the FinishForeignModify function.
That is, on errors, the context is not released, including
curl-connections to the Java-part are not closed.
Therefore, I added a callback to release resources on errors,
which releases the context, including closing curl-connections.