Incorrect resource release callback behavior in b244843
If during an error the resource release callback was called,
and some error occurs in it, then the transaction abort is called again.
If the callback was not removed from the list of callbacks,
then it is called again. If an error occurs again,
the transaction is aborted again and the callback is called again.
And this happens until the error stack is exhausted,
after which panic occurs.

So the solution is two steps:

1) at the very beginning of the execution of the resource release callback,
remove it from the list of callbacks so that it is no longer called

2) when processing the callback, intercept all errors,
turning them into warnings