Revert "Refactor closing connection to pxf external-table and fdw"
Recent changes to PXF provides multiple bugs in churl_cleanup:

1) If during an error the resource release callback was called,
and some error occurs in it, then the transaction abort is called again.
If the callback was not removed from the list of callbacks,
then it is called again. If an error occurs again, the transaction
is aborted again and the callback is called again. And this happens until
the error stack is exhausted, after which panic occurs.

2) Buffers free was removed. It caller has init->cleanup cycle inside
the same context, it cause memory leak. Current (fixed) case is a special case,
when buffers are always freed with context.
In general case there will be a memory leak.

3) Headers free is now done on churl level. It breaks the contract,
because headers are allocated somewhere outside churl (in bridge in our case).
In general case the caller expects that headers are valid after churl_cleanup().
More, in case of fdw they are allocated in child portal context (ExecutorState),
which is freed before calling callbacks in case of transaction abort.

This reverts commit b244843.