Canceling read requests in the PXF service
In case of any errors, including request cancellation, the patch
2fd4adc closes curl connections to the PXF connector.
However, for read requests, the PXF connector
is not notified of this event because it is waiting for blocked I/O.
For example, it waits for a remote JDBC request to complete.
This wastes resources on the PXF connector and on the remote side
because the query results are no longer needed.

This patch enhances the functionality of the PXF protocol in clearing
the execution context when a read request fails. In this case, it calls
the POST /pxf/cancel endpoint on the PXF connector with the same headers
as the main request, adding a header with the client port.

This patch works with the Java part of the patch fbfff8c.