Cancel read requests in PXF service
The C-part of the PXF external table releases the context
(cleanup_context -> gpbridge_cleanup) only on the last call
in the pxfprotocol_export and pxfprotocol_import functions.
The C-part of the PXF FDW releases the context (PxfBridgeCleanup)
only in the FinishForeignModify function.
That is, on errors, the context is not released, including
curl-connections to the Java-part are not closed.

In case of any errors, for read requests, the PXF connector
is not notified of this event because it is waiting for blocked I/O.
For example, it waits for a remote JDBC request to complete.
This wastes resources on the PXF connector and on the remote side
because the query results are no longer needed.

Therefore, I added a callback to release resources on errors,
which releases the context, including closing curl-connections.
Also, this patch enhances the functionality of the PXF protocol in clearing
the execution context when a read request fails. In this case, it calls
the POST /pxf/cancel endpoint on the PXF connector with the same
headers as the main request, adding a header with the client port
and 1 second timeout.

This patch works with the Java part of the patch fbfff8c.