Throw error on unsupported operator for json[]
Due to GPDB do not support any scalar operator within json
type, we throw the error for the case of filter with json[]
instead of letting executor send the read request to pxf without
filter. Otherwise the executor will eventually cut off the query
and throw the error that filter operator is unsupported after
the data has been processed and returned by the pxf. In order
to avoid extra work, we throw an error before the request to pxf.